requirements.txt
# base 
yolov5
torchreid
strongsort
